requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
openai==1.14.2
PyMuPDF==1.23.14
pdfplumber==0.10.3
python-docx==1.1.0
pytesseract==0.3.10
Pillow==10.2.0
faiss-cpu==1.12.0
numpy==1.26.4
tiktoken==0.5.2
nltk==3.8.1
Werkzeug==2.3.7
mysql-connector-python==8.1.0
PyJWT==2.8.0
bcrypt==4.0.1